+++ src/pydapter/fields/__init__.py
@@ -1,3 +1,27 @@
+from pydapter.fields.builder import DomainModelBuilder
+from pydapter.fields.common_templates import (
+    CREATED_AT_TEMPLATE,
+    CREATED_AT_TZ_TEMPLATE,
+    DELETED_AT_TEMPLATE,
+    DELETED_AT_TZ_TEMPLATE,
+    DESCRIPTION_TEMPLATE,
+    EMAIL_TEMPLATE,
+    ID_TEMPLATE,
+    JSON_TEMPLATE,
+    METADATA_TEMPLATE,
+    NAME_TEMPLATE,
+    NONNEGATIVE_INT_TEMPLATE,
+    PERCENTAGE_TEMPLATE,
+    PHONE_TEMPLATE,
+    POSITIVE_FLOAT_TEMPLATE,
+    POSITIVE_INT_TEMPLATE,
+    STRING_TEMPLATE,
+    TAGS_TEMPLATE,
+    UPDATED_AT_TEMPLATE,
+    UPDATED_AT_TZ_TEMPLATE,
+    URL_TEMPLATE,
+    USERNAME_TEMPLATE,
+)
 from pydapter.fields.dts import (
     DATETIME,
     DATETIME_NULLABLE,
@@ -6,6 +30,7 @@ from pydapter.fields.dts import (
 )
 from pydapter.fields.embedding import EMBEDDING, validate_embedding
 from pydapter.fields.execution import EXECUTION, Execution
+from pydapter.fields.families import FieldFamilies, create_field_dict
 from pydapter.fields.ids import (
     ID_FROZEN,
     ID_MUTABLE,
@@ -20,6 +45,11 @@ from pydapter.fields.params import (
     validate_model_to_params,
     validate_model_to_type,
 )
+from pydapter.fields.protocol_families import (
+    ProtocolFieldFamilies,
+    create_protocol_model,
+)
+from pydapter.fields.template import FieldTemplate
 from pydapter.fields.types import (
     ID,
     Embedding,
@@ -28,6 +58,11 @@ from pydapter.fields.types import (
     Undefined,
     UndefinedType,
     create_model,
+)
+from pydapter.fields.validation_patterns import (
+    ValidationPatterns,
+    create_pattern_template,
+    create_range_template,
 )
 
 __all__ = (
@@ -56,4 +91,34 @@ __all__ = (
     "ID",
     "Embedding",
     "Metadata",
+    "FieldTemplate",
+    "ID_TEMPLATE",
+    "STRING_TEMPLATE",
+    "EMAIL_TEMPLATE",
+    "USERNAME_TEMPLATE",
+    "CREATED_AT_TEMPLATE",
+    "UPDATED_AT_TEMPLATE",
+    "DELETED_AT_TEMPLATE",
+    "CREATED_AT_TZ_TEMPLATE",
+    "UPDATED_AT_TZ_TEMPLATE",
+    "DELETED_AT_TZ_TEMPLATE",
+    "NAME_TEMPLATE",
+    "DESCRIPTION_TEMPLATE",
+    "URL_TEMPLATE",
+    "PHONE_TEMPLATE",
+    "POSITIVE_INT_TEMPLATE",
+    "NONNEGATIVE_INT_TEMPLATE",
+    "POSITIVE_FLOAT_TEMPLATE",
+    "PERCENTAGE_TEMPLATE",
+    "JSON_TEMPLATE",
+    "TAGS_TEMPLATE",
+    "METADATA_TEMPLATE",
+    "FieldFamilies",
+    "create_field_dict",
+    "DomainModelBuilder",
+    "ProtocolFieldFamilies",
+    "create_protocol_model",
+    "ValidationPatterns",
+    "create_pattern_template",
+    "create_range_template",
 )